♻️ refactor: Rename ConfigProvider to Config and streamline context structure
- Move the providers ConfigProvider directory to Config and change the default export to Config
- Split ConfigContext into a separate context file and rename the type to ContextValue
- Update providers index exports to use Config and ConfigProps
- Replace ConfigProvider usages with Config across docs stories
- Add a commit message generation skill document to define the commit writing workflow
