+++ apps/docs/stories/atoms/Config/index.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
-import { Button, ConfigProvider, Switch } from '@repo/ui';
+import { Button, Config, Switch } from '@repo/ui';
 import { cn } from '@repo/ui/utils';
 
-const meta: Meta<typeof ConfigProvider> = {
-  title: 'UI/ConfigProvider',
-  component: ConfigProvider,
+const meta: Meta<typeof Config> = {
+  title: 'UI/Config',
+  component: Config,
   parameters: {
     layout: 'centered',
   },
@@ -22,7 +22,7 @@ export const ThemeToken: Story = {
         <Switch />
       </div>
 
-      <ConfigProvider
+      <Config
         theme={{
           token: { colorPrimary: 'oklch(0.6 0.25 30)' },
         }}
@@ -31,9 +31,9 @@ export const ThemeToken: Story = {
           <Button>Red Primary</Button>
           <Switch />
         </div>
-      </ConfigProvider>
+      </Config>
 
-      <ConfigProvider
+      <Config
         theme={{
           token: { colorPrimary: 'oklch(0.55 0.2 260)' },
         }}
@@ -42,7 +42,7 @@ export const ThemeToken: Story = {
           <Button>Blue Primary</Button>
           <Switch />
         </div>
-      </ConfigProvider>
+      </Config>
     </div>
   ),
 };
@@ -55,36 +55,34 @@ export const DarkMode: Story = {
         <Button>Button</Button>
       </div>
 
-      <ConfigProvider theme={{ dark: true }}>
+      <Config theme={{ dark: true }}>
         <div
           className={cn('bg-background text-foreground rounded-lg border p-4')}
         >
           <p className={cn('mb-2 text-sm')}>Dark</p>
           <Button>Button</Button>
         </div>
-      </ConfigProvider>
+      </Config>
     </div>
   ),
 };
 
 export const Nested: Story = {
   render: () => (
-    <ConfigProvider theme={{ token: { colorPrimary: 'oklch(0.55 0.2 260)' } }}>
+    <Config theme={{ token: { colorPrimary: 'oklch(0.55 0.2 260)' } }}>
       <div className={cn('flex flex-col gap-4')}>
         <div className={cn('flex items-center gap-2')}>
           <Button>Blue (Parent)</Button>
           <Switch />
         </div>
 
-        <ConfigProvider
-          theme={{ token: { colorPrimary: 'oklch(0.6 0.2 150)' } }}
-        >
+        <Config theme={{ token: { colorPrimary: 'oklch(0.6 0.2 150)' } }}>
           <div className={cn('flex items-center gap-2')}>
             <Button>Green (Child)</Button>
             <Switch />
           </div>
-        </ConfigProvider>
+        </Config>
       </div>
-    </ConfigProvider>
+    </Config>
   ),
 };
