--- packages/ui/src/providers/ConfigProvider/ConfigContext.ts
@@ -1,11 +1,0 @@
-'use client';
-
-import { createContext, useContext } from 'react';
-
-import type { ConfigContextValue } from './types';
-
-export const ConfigContext = createContext<ConfigContextValue>({
-  theme: {},
-});
-
-export const useConfig = () => useContext(ConfigContext);
